[#1222] Split resources per fiscal year
As the final step in the pipeline, before uploading the data to OpenSpending,
we go over the datapackage looking for a column with
`osType == 'date:fiscal-year'` (there can only be one). If there's no such
column, the datapackage isn't modified, if there is, we go over the data and
create one resource per year.

The pre-conditions for this processor are:

* There is only (and only one) column with `osType == 'date:fiscal-year'`
* There is only one resource

If either pre-condition isn't met, we return the datapackage unmodified.

Fixes openspending/openspending#1222