[#1222] Fix bug with resources outside the DP's base path
Currently, `datapackage-py` only supports "safe" resources, which when they're
local resources it means that they must be located somewhere inside the
datapackage.json's directory.

I tried setting the DataPackage's base path as the directory where temporary
files are created (e.g. `/tmp`), but that didn't work because the DP already
has other local resources that don't live in `/tmp`. As the DataPackage that
I'm loading is from a zip file, the directory its data live in exist only in
memory, so I can't add new files to it.

In the end, the cleanest solution was monkey patching (yeah, I know...). I
simply mock the `datapackage.helpers.is_safe_path()` to always return true when
saving the DataPackage, and everything works.

This is a temporary solution. I created an issue on
https://github.com/frictionlessdata/datapackage-py/issues/171 to find a way
that won't require monkey patching.

openspending/openspending#1222