[#1222] Split tasks into multiple pipelines
We were having issues with the subsequent tasks running while the previous ones
were still working. This is a design decision of `datapackage-pipelines`: all
tasks run concurrently, so we get better performance.

We used to do a hack to wait for the previous tasks to complete (by passing on
all received data before executing), but multiple pipelines are cleaner, and
hopefully easier to debug.

openspending/openspending#1222